% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bqrDE}
\alias{bqrDE}
\title{BQR Based on DE}
\usage{
bqrDE(Xr, Yr, p, nsim, initCoef, burn = 5000L)
}
\arguments{
\item{Xr}{Matrix of covariate.}

\item{Yr}{Vector of response.}

\item{p}{Quantile of interest.}

\item{nsim}{Number of simulation.}

\item{initCoef}{Initial value (vector) of the model coefficients.}

\item{burn}{Number of burn-in.}
}
\value{
List of model coefficients and acceptance rate.
}
\description{
Bayesian quantile regression using double exponential distribution (DE) as Likelihood Function.
}
