% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bqrDPML}
\alias{bqrDPML}
\title{BQR Based on DPM of Logistic Distribution}
\usage{
bqrDPML(Xr, Yr, p, nsim, initCoef, sampleQ = TRUE, burn = 5000L,
  plotDensity = FALSE, naux = 10L, threshold = 1e-07)
}
\arguments{
\item{Xr}{Matrix of covariate.}

\item{Yr}{Vector of response.}

\item{p}{Quantile of interest.}

\item{nsim}{Number of simulation.}

\item{initCoef}{Initial value (vector) of the model coefficients.}

\item{sampleQ}{A boolean variable which determines how to calculate the location shift.
(true means using sample quantile, false means using the density estimation)}

\item{burn}{Number of burn-in.}

\item{plotDensity}{A boolean variable which decides whether to plot the density estimation or not.}

\item{naux}{Number of the auxiliary variable.}

\item{threshold}{The threshold of the error of the solution to the equation of the lambda.}
}
\value{
List of model parameters including the coefficients.
}
\description{
Bayesian quantile regression with Dirichlet process mixture of Logistic Distribution.
It is the main and recommended method of this package.
}
