% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bqrDPMN}
\alias{bqrDPMN}
\title{BQR based on DPM of Normal Distribution}
\usage{
bqrDPMN(Xr, Yr, p, nsim, initCoef, sampleQ = TRUE, burn = 5000L,
  plotDensity = FALSE)
}
\arguments{
\item{Xr}{Matrix of covariate.}

\item{Yr}{Vector of response.}

\item{p}{Quantile of interest.}

\item{nsim}{Number of simulation.}

\item{initCoef}{Initial value (vector) of the model coefficients.}

\item{sampleQ}{A boolean variable which determines how to calculate the location shift.
(true means using sample quantile, false means using the density estimation)}

\item{burn}{Number of burn-in.}

\item{plotDensity}{A boolean variable which decides whether to plot the density estimation or not.}
}
\value{
List of model parameters including the coefficients.
}
\description{
Bayesian quantile regression with Dirichlet process mixture of normal distribution.
}
