% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bqrDPMNHP}
\alias{bqrDPMNHP}
\title{BQR based on DPM of Normal Distribution? (what is 'P'?)}
\usage{
bqrDPMNHP(Xr, Yr, T, p, nsim, initCoef, burn = 5000L, plotDensity = FALSE,
  hasHeter = TRUE)
}
\arguments{
\item{Xr}{Matrix of covariate.}

\item{Yr}{Vector of response.}

\item{p}{Quantile of interest.}

\item{nsim}{Number of simulation.}

\item{initCoef}{Initial value (vector) of the model coefficients.}

\item{burn}{Number of burn-in.}

\item{plotDensity}{A boolean variable which decides whether to plot the density estimation or not.}

\item{hasHeter}{Boolean indicating if Heteroskedasticity is present. Default is True.}

\item{sampleQ}{A boolean variable which determines how to calculate the location shift.
(true means using sample quantile, false means using the density estimation)}
}
\value{
List of model parameters including the coefficients.
}
\description{
Bayesian quantile regression with heteroskedasticity with Dirichlet process mixture of normal distribution.
}
