% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bqrDPMU}
\alias{bqrDPMU}
\title{BQR based on DPM of Uniform Distribution}
\usage{
bqrDPMU(Xr, Yr, p, nsim, initCoef, burn = 5000L, plotDensity = FALSE,
  tol = 1e-07)
}
\arguments{
\item{Xr}{Matrix of covariate.}

\item{Yr}{Vector of response.}

\item{p}{Quantile of interest.}

\item{nsim}{Number of simulation.}

\item{initCoef}{Initial value (vector) of the model coefficients.}

\item{burn}{Number of burn-in.}

\item{plotDensity}{A boolean variable which decides whether to plot the density estimation or not.}

\item{tol}{The tolerance value for the length of the interval to treat the normal density in this interval as constant.}
}
\value{
List of model parameters including the coefficients.
}
\description{
Bayesian quantile regression with Dirichlet process mixture of uniform distribution.
}
