% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bqrDPMUH}
\alias{bqrDPMUH}
\title{BQR with Heteroskedasticity based on DPM of Uniform Distribution}
\usage{
bqrDPMUH(Xr, Yr, p, nsim, initCoef, burn = 5000L, plotDensity = FALSE,
  tol = 1e-07, hasHeter = TRUE)
}
\arguments{
\item{Xr}{Matrix of covariate.}

\item{Yr}{Vector of response.}

\item{p}{Quantile of interest.}

\item{nsim}{Number of simulation.}

\item{initCoef}{Initial value (vector) of the model coefficients.}

\item{burn}{Number of burn-in.}

\item{plotDensity}{A boolean variable which decides whether to plot the density estimation or not.}

\item{tol}{The tolerance value for the length of the interval to treat the normal density in this interval as constant.}

\item{hasHeter}{Boolean indicating if Heteroskedasticity is present. Default is True.}
}
\value{
List of model parameters including the coefficients.
}
\description{
Bayesian quantile regression with Heteroskedasticity based on Dirichlet process mixture of uniform distribution.
}
