% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bqrNH}
\alias{bqrNH}
\title{BQR with Heteroscedasticity Based on Normal Distribution}
\usage{
bqrNH(Xr, Yr, p, nsim, initCoef, burn = 5000L)
}
\arguments{
\item{Xr}{Matrix of covariate.}

\item{Yr}{Vector of response.}

\item{p}{Quantile of interest.}

\item{nsim}{Number of simulation.}

\item{initCoef}{Initial value (vector) of the model coefficients.}

\item{burn}{Number of burn-in.}
}
\value{
List of model coefficients and acceptance rate.
}
\description{
Bayesian quantile regression with heteroscedasticity using normal distribution as Likelihood Function.
}
