% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bqrNM}
\alias{bqrNM}
\title{BQR Based on Mixture of Normal Distribution}
\usage{
bqrNM(Xr, Yr, p, nsim, initCoef, weight, mu, sigma, burn = 5000L)
}
\arguments{
\item{Xr}{Matrix of covariate.}

\item{Yr}{Vector of response.}

\item{p}{Quantile of interest.}

\item{nsim}{Number of simulation.}

\item{initCoef}{Initial value (vector) of the model coefficients.}

\item{burn}{Number of burn-in.}
}
\value{
List of model coefficients and acceptance rate.
}
\description{
Bayesian quantile regression using a mixture of normal distributions as Likelihood Function.
}
