% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bqrT3}
\alias{bqrT3}
\title{BQR with based on t Distribution}
\usage{
bqrT3(Xr, Yr, p, nsim, initCoef, burn = 5000L)
}
\arguments{
\item{Xr}{Matrix of covariate.}

\item{Yr}{Vector of response.}

\item{p}{Quantile of interest.}

\item{nsim}{Number of simulation.}

\item{initCoef}{Initial value (vector) of the model coefficients.}
}
\value{
List of model parameters including the coefficients.
}
\description{
Bayesian quantile regression using the t-distribution with Degree of Freedom 3 as Likelihood Function.
}
