% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{checkLoss}
\alias{checkLoss}
\title{The Check Loss Function}
\usage{
checkLoss(y, yHat, tau)
}
\arguments{
\item{y}{Vector of observed response.}

\item{yHat}{Vector of fitted response.}

\item{tau}{Quantile of interest.}
}
\value{
check loss between fitted and observed responses.
}
\description{
Check loss function for frequentist quantile regression.
}
