% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleQuantile}
\alias{sampleQuantile}
\title{Sample Quantile Function}
\usage{
sampleQuantile(array, p)
}
\arguments{
\item{array}{Vector of sample points.}

\item{p}{Quantile of interest.}
}
\value{
the specified sample quantile of the given sample.
}
\description{
Compute the specified sample quantile of a given sample.
}
